+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.0.1
 
-* A Flutter package that provides a customizable DottedSeparator widget that can be used to create a dotted line separator with adjustable color and dash width.
+* A Flutter package that provides a customizable Dotted Separator widget that can be used to create a dotted line separator with adjustable color and dash width.
 
 ## 0.0.2
 
@@ -13,3 +13,7 @@
 ## 1.0.1
 
 * Updated screenshot.
+
+## 1.0.2
+
+* Added decoration options for the dotted widget.